fix same barrel mistakes, add home screen

+++ lib/core/constants/app_sizes.dart
@@ -1,24 +1,24 @@
 class AppSizes {
-  static const int defaultScreenWidth = 375;
-  static const int defaultScreenHeight = 812;
-  static const int defaultAuthScreenHorizontalPadding = 20;
-  static const int defaultHomeScreenHorizontalPadding = 24;
-  static const int defaultPaintScreenHorizontalPadding = 21;
-  static const int defaultFormInputContentPadding = 16;
-  static const int defaultBorderRadius = 8;
-  static const int defaultFormInputHeight = 78;
-  static const int defaultFormGap = 20;
-  static const int defaultHomeImageWrapSpacing = 15;
-  static const int defaultBtnHeight = 48;
-  static const int defaultIconBtnSize = 38;
-  static const int defaultImageCardBorderRadius = 12;
+  static const double defaultScreenWidth = 375;
+  static const double defaultScreenHeight = 812;
+  static const double defaultAuthScreenHorizontalPadding = 20;
+  static const double defaultHomeScreenHorizontalPadding = 24;
+  static const double defaultPaintScreenHorizontalPadding = 21;
+  static const double defaultFormInputContentPadding = 16;
+  static const double defaultBorderRadius = 8;
+  static const double defaultFormInputHeight = 78;
+  static const double defaultFormGap = 20;
+  static const double defaultHomeImageWrapSpacing = 15;
+  static const double defaultBtnHeight = 48;
+  static const double defaultIconBtnSize = 38;
+  static const double defaultImageCardBorderRadius = 12;
   static const double defaultImageCardWidth =
       (defaultScreenWidth - 2 * defaultHomeScreenHorizontalPadding - defaultHomeImageWrapSpacing) /
       2;
-  static const int defaultImageEditableHeight = 510;
+  static const double defaultImageEditableHeight = 510;
   static const double fontSize20 = 20;
   static const double fontSize17 = 17;
   static const double fontSize12 = 12;
   static const double fontSize14 = 14;
-  static const int defaultIconSize = 24;
+  static const double defaultIconSize = 24;
 }
